have to donation positive amount, show last 5 donations on main page, can simulate making a purchase with myCoins

+++ app/src/main/java/com/example/wegive/User.kt
@@ -4,7 +4,7 @@ import com.google.firebase.firestore.PropertyName
 
 //
 data class User(var userName: String = "", var email: String= "", var firstName: String= "", var lastName: String= "",
-                var totalAmountGiven: Int=0, var myCoins: Int=0, var totalNumberOfDonations: Int=0, var useForColu: Boolean=true,
+                var totalAmountGiven: Int=0, var myCoins: Double= 0.0, var totalNumberOfDonations: Int=0, var useForColu: Boolean=true,
                 @get:PropertyName("profile_image_url") @set: PropertyName("profile_image_url") var profile_image_url: String = "", var hasCC: Boolean = false, var last4: String=""){
     override fun toString(): String {
         return "$userName $firstName $lastName"
